writing-practice/requirements.txt
groq
dotenv
requests
pytesseract 
Pillow opencv-python